# api-docs.json
{
  "openapi": "3.1.0",
  "info": { "title": "DevSolve" },
  "servers": [
    {
      "url": "https://devsolve-api.quizzy.it.com",
      "description": "Generated server url"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/follows/{type}/{targetId}": {
      "put": {
        "tags": ["follow-controller"],
        "operationId": "follow",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["ORGANIZATION", "USER", "PROBLEM", "PROGRAM", "SHOWCASE"]
            }
          },
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/FollowResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["follow-controller"],
        "operationId": "unfollow",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["ORGANIZATION", "USER", "PROBLEM", "PROGRAM", "SHOWCASE"]
            }
          },
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": { "204": { "description": "No Content" } }
      }
    },
    "/api/v1/categories/{id}/icon": {
      "put": {
        "tags": ["category-controller"],
        "operationId": "uploadIcon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CategoryResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["category-controller"],
        "operationId": "removeIcon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CategoryResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organizations/me/members/{userId}/permissions": {
      "patch": {
        "tags": ["organization-controller"],
        "operationId": "updateMemberPermissions",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberPermissionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/MemberResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/solutions/{id}/review-status": {
      "patch": {
        "tags": ["solution-admin-controller"],
        "operationId": "updateReviewStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolutionReviewStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/SolutionResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/solutions/{id}": {
      "get": {
        "tags": ["solution-admin-controller"],
        "operationId": "getReviewDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/SolutionResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/organizations/pending": {
      "get": {
        "tags": ["organization-admin-controller"],
        "operationId": "getPendingOrganizations",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageOrganizationReviewSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FollowResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "followableType": {
            "type": "string",
            "enum": ["ORGANIZATION", "USER", "PROBLEM", "PROGRAM", "SHOWCASE"]
          },
          "followableId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "scope": { "type": "string", "enum": ["PROBLEM", "SHOWCASE"] },
          "description": { "type": "string" },
          "iconUrl": { "type": "string" },
          "sortOrder": { "type": "integer", "format": "int32" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "SolutionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "problemId": { "type": "string", "format": "uuid" },
          "authorId": { "type": "string", "format": "uuid" },
          "description": { "type": "string" },
          "videoUrl": { "type": "string" },
          "diagramUrl": { "type": "string" },
          "reviewStatus": {
            "type": "string",
            "enum": ["PENDING", "APPROVED", "REJECTED", "ACCEPTED"]
          },
          "reviewedBy": { "type": "string", "format": "uuid" },
          "reviewedAt": { "type": "string", "format": "date-time" },
          "rejectionReason": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "MemberResponse": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["MANAGER", "MEMBER", "VIEWER"] },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "VIEW_PROGRAMS",
                "CREATE_PROGRAM",
                "EDIT_PROGRAM",
                "MANAGE_PROGRAM_STATE",
                "VIEW_REPORTS",
                "TRIAGE_REPORTS",
                "MANAGE_DISCLOSURE",
                "AWARD_REWARDS"
              ]
            },
            "uniqueItems": true
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED", "REMOVED"]
          },
          "invitationPending": { "type": "boolean" },
          "joinedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "UpdateMemberPermissionsRequest": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "VIEW_PROGRAMS",
                "CREATE_PROGRAM",
                "EDIT_PROGRAM",
                "MANAGE_PROGRAM_STATE",
                "VIEW_REPORTS",
                "TRIAGE_REPORTS",
                "MANAGE_DISCLOSURE",
                "AWARD_REWARDS"
              ]
            },
            "maxItems": 8,
            "minItems": 0,
            "uniqueItems": true
          }
        },
        "required": ["permissions"]
      },
...
      "UpdateSolutionReviewStatusRequest": {
        "type": "object",
        "properties": {
          "reviewStatus": {
            "type": "string",
            "enum": ["PENDING", "APPROVED", "REJECTED", "ACCEPTED"]
          },
          "rejectionReason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          }
        },
        "required": ["reviewStatus"]
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": { "type": "integer", "format": "int64" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "unpaged": { "type": "boolean" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "pageSize": { "type": "integer", "format": "int32" },
          "paged": { "type": "boolean" }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": { "type": "boolean" },
          "sorted": { "type": "boolean" },
          "unsorted": { "type": "boolean" }
        }
      },
...
      "OrganizationReviewSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "websiteUrl": { "type": "string" },
          "industry": {
            "type": "string",
            "enum": [
              "TECHNOLOGY",
              "FINANCE",
              "HEALTHCARE",
              "ECOMMERCE",
              "GOVERNMENT",
              "EDUCATION",
              "OTHER"
            ]
          },
          "companySize": { "type": "string" },
          "country": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACTIVE", "REJECTED"]
          },
          "ownerId": { "type": "string", "format": "uuid" },
          "ownerFullName": { "type": "string" },
          "ownerEmail": { "type": "string" },
          "submissionVersion": { "type": "integer", "format": "int32" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "PageOrganizationReviewSummaryResponse": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationReviewSummaryResponse"
            }
          },
          "number": { "type": "integer", "format": "int32" },
          "numberOfElements": { "type": "integer", "format": "int32" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "pageable": { "$ref": "#/components/schemas/PageableObject" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "empty": { "type": "boolean" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}